All projects now target .NET framework 4.6.1. Updated package reference for PdfSharp to version 1.50.4619-beta4c.
Added small change in FontsHandler so that when a font cannot be fond in the existing fonts the adapter will be given a chance to locate it. In the case of PdfSharp this allows using an  implementation of the IFontResolver interface which in turn allows the use of custom fonts when generating PDF file. Added a demo console application HtmlRenderer.Demo.PdfSharp which shows how to convert HTML to PDF including the use of a custom font via a custom FontResolver class.

Also updated the WinForms example. For it's button to generate a PDF File, it now also makes use of a custom font.

Replaced packages.config with PackageReference in the project files in some cases. Solution builds with Visual Studio 2017 version 15.3.5.